Refactor: Organize packages
